upgrade to juntagrico 1.6.0 [wip]
