+++ juntagrico_custom_sub/apps.py
@@ -12,3 +12,8 @@ class JuntagricoCustomSubAppconfig(AppConfig):
         # for compatibility to downgraded subscription views
         addons.config.register_sub_overview('cs/hooks/subscription_content_overview.html')
         addons.config.register_sub_change('cs/hooks/subscription_content_change.html')
+
+        # Basimilch specific
+        from juntagrico.forms import SubscriptionPartBaseForm
+        from juntagrico_custom_sub.basimilch import quantity_error
+        SubscriptionPartBaseForm.validators.append(quantity_error)
