+++ juntagrico_custom_sub/templates/cs/content_edit_result.html
@@ -8,8 +8,10 @@
 
 {% block content %}
     <div class="alert alert-success" role="alert">
-        Der Inhalt deines {% vocabulary "subscription" %}s wurde gespeichert. <br/> <a href='/my/subscription/detail/'>
-        Zurück zur {% vocabulary "subscription" %}-Übersicht </a>
+        Der Inhalt deines {% vocabulary "subscription" %}s wurde gespeichert. <br/>
+        <a href="{% url 'subscription-landing' %}">
+            Zurück zur {% vocabulary "subscription" %}-Übersicht
+        </a>
     </div>
 {% endblock %}
