--- juntagrico_custom_sub/templates/cs/initial_select_size.html
@@ -1,5 +1,0 @@
-{% extends "createsubscription/select_subscription.html" %}
-{% block all_page_title %}
-    {% include "cs/snippets/error_message.html" with initial=True %}
-    {{ block.super }}
-{% endblock %}
