--- juntagrico_custom_sub/templates/cs/initial_summary.html
@@ -1,108 +1,0 @@
-{% extends "no_menu_base.html" %}
-{% load i18n %}
-{% load juntagrico.config %}
-{% load crispy_forms_tags %}
-{% block all_page_title %}
-    {% vocabulary "subscription" as v_subscription %}
-    <div class="offset-md-3 col-md-9 juntagrico">
-        <h3>
-            {% blocktrans %}Zusammenfassung{% endblocktrans %}
-        </h3>
-    </div>
-{% endblock %}
-{% block allcontent %}
-    {% config "organisation_name" as c_organisation_name %}
-    {% config "share_price" as c_share_price %}
-    {% config "currency" as c_currency %}
-    {% vocabulary "member_type" as v_member_type %}
-    {% vocabulary "member_type_pl" as v_member_type_pl %}
-    {% vocabulary "co_member_pl" as v_co_member_pl %}
-    {% vocabulary "subscription" as v_subscription %}
-    {% vocabulary "subscription_pl" as v_subscription_pl %}
-    {% config "enable_shares" as c_enable_shares %}
-    <div class="offset-md-2 col-md-8">
-        <h4>{% trans "Personalien" %}
-            {% if not main_member.pk %}
-                <a href="{% url 'signup' %}" class="edit"><i class="fas fa-pen"></i></a>{% endif %}
-        </h4>
-        <p>{{ main_member.first_name }} {{ main_member.last_name }} <br>{{ main_member.addr_street }}
-            <br>{{ main_member.addr_zipcode }} {{ main_member.addr_location }} <br>{% trans 'E-Mail' %}:
-            <b>{{ main_member.email }}</b> <br>{% trans 'Telefon' %}: {{ main_member.phone }}
-            <br>{% trans 'Mobil' %}: {{ main_member.mobile_phone|default:'-' }}
-            <br>{% trans 'Geburtstag' %}: {{ main_member.birthday|default:'-' }}
-        </p>
-
-        <h4>{{ v_subscription }} <a href="{% url 'cs-subscription' %}" class="edit"><i class="fas fa-pen"></i></a></h4>
-        <p>
-            {% for type, amount in subscriptions.items %}
-                {% blocktrans trimmed with tn=type.name tp=type.price %}
-                    <b>{{ amount }}</b>&times; {{ tn }} à {{ c_currency }} {{ tp }}/Jahr
-                {% endblocktrans %}
-                {% if not forloop.last %}<br>{% endif %}
-            {% endfor %}
-            {% if subscriptions|length == 0 %}
-                {% blocktrans %}Keine {{ v_subscription_pl }}{% endblocktrans %}
-            {% endif %}
-        </p>
-
-        <h4>{% trans "Inhalt" %} <a href="{% url 'custom_sub_initial_select' %}" class="edit"><i class="fas fa-pen"></i></a>
-        </h4>
-        {% for product, amount in custom_prod.items %}
-            <b>{{ amount }}</b>&times; {{ product|striptags }} <br>
-        {% endfor %}
-
-        {% if subscriptions|length > 0 %}
-            <h4>{% trans "Depot" %} <a href="{% url 'cs-depot' %}" class="edit"><i class="fas fa-pen"></i></a></h4>
-            <p>{{ depot.name }} <br>{% trans 'Abholung jeweils am' %} {{ depot.weekday_name }}
-                {% if depot.addr_street %}<br>{{ depot.addr_street }}{% endif %}
-                {% if depot.addr_zipcode or depot.addr_location %}<br>{{ depot.addr_zipcode }}
-                    {{ depot.addr_location }}{% endif %}
-            </p>
-
-            <h4>{% trans "Gewünschtes Startdatum" %} <a href="{% url 'cs-start' %}" class="edit"><i
-                    class="fas fa-pen"></i></a></h4>
-            <p>{{ start_date }}</p>
-
-            <h4>{{ v_co_member_pl }}</h4>
-            {% for co_member in co_members %}
-                <p>{{ co_member.first_name }} {{ co_member.last_name }}
-                    {% if not co_member.pk %}
-                        <a href="{% url 'cs-co-members' %}?edit={{ forloop.counter }}" class="edit"><i
-                                class="fas fa-pen"></i></a>{% endif %}
-                    <a href="{% url 'cs-co-members' %}?remove={{ forloop.counter }}" class="delete"><i
-                            class="fas fa-trash"></i></a> <br>{{ co_member.addr_street }}
-                    <br>{{ co_member.addr_zipcode }} {{ co_member.addr_location }} <br>{% trans 'E-Mail' %}:
-                    <b>{{ co_member.email }}</b> <br>{% trans 'Telefon' %}: {{ co_member.phone }}
-                    <br>{% trans 'Mobil' %}: {{ co_member.mobile_phone|default:'-' }}
-                    <br>{% trans 'Geburtstag' %}: {{ co_member.birthday|default:'-' }}
-                </p>
-            {% endfor %}
-            <p><a href="{% url 'cs-co-members' %}"><i class="fas fa-plus"></i>
-                {% if co_members|length > 0 %}
-                    {% blocktrans %}Weitere {{ v_co_member_pl }} hinzufügen{% endblocktrans %}
-                {% else %}
-                    {% blocktrans %}{{ v_co_member_pl }} hinzufügen{% endblocktrans %}
-                {% endif %}
-            </a></p>
-        {% endif %}
-
-        {% if c_enable_shares %}
-            <h4>{% trans "Anteilscheine" %} <a href="{% url 'cs-shares' %}" class="edit"><i class="fas fa-pen"></i></a>
-            </h4>
-            <p>
-                {{ main_member.first_name }} {{ main_member.last_name }}:
-                <b>{{ main_member.new_shares|default:'Keine' }}</b>
-                {% for co_member in co_members %}
-                    <br>{{ co_member.first_name }} {{ co_member.last_name }}:
-                    <b>{{ co_member.new_shares|default:'Keine' }}</b>
-                {% endfor %}
-            </p>
-            <p>{% blocktrans %}Zu je einmalig {{ c_currency }} {{ c_share_price }}{% endblocktrans %}</p>
-        {% endif %}
-
-        <h4>Kommentar</h4>
-        <form action="" method="post">
-            {% crispy form %}
-        </form>
-    </div>
-{% endblock %}
