--- juntagrico_custom_sub/templates/cs/snippets/error_message.html
@@ -1,10 +1,0 @@
-{% load i18n %}
-{% load juntagrico.config %}
-{% if error %}
-    {% if initial %}
-        <div class="col-md-8 offset-2"> {% endif %}
-<div class="alert alert-danger" role="alert">
-    {{ error }}
-</div>
-{% if initial %} </div> {% endif %}
-{% endif %}
