+++ juntagrico_custom_sub/templates/juntagrico/my/subscription/single.html
@@ -1,0 +1,5 @@
+{% extends "juntagrico/my/subscription/single.html" %}
+{% block part-list %}
+    {{ block.super }}
+    {% include "cs/snippets/content_overview.html" %}
+{% endblock %}
