+++ pyproject.toml
@@ -12,7 +12,7 @@ readme = "README.md"
 license = { file = "LICENSE" }
 requires-python = ">=3.8"
 dependencies = [
-    "juntagrico==1.6.0rc4",
+    "juntagrico @ git+https://github.com/juntagrico/juntagrico.git@release_1.6",
 ]
 classifiers = [
     'Development Status :: 5 - Production/Stable',
